refactor: js files
